added sidebar
