+++ src/components/post/sidebar/postview.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const PostView = () => {
+  return (
+<div>      
+    </div>
+  )
+}
+
+export default PostView
